pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Production-grade extensible autonomous FX + multi-tool AI agent"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "openai>=1.60",
    "mistralai>=1.0",
    "requests",
    "python-dotenv",
    "pytest",
    "pytest-cov",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.mypy]
strict = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra --cov=src --cov-report=term-missing"
testpaths = ["tests"]